# src/main/java/com/example/flightreservation/models/responses/FlightResponse.java
package com.example.flightreservation.models.responses;

public class FlightResponse {
    private Integer flightId;
    private String flightNumber;
    private Integer status;

    public Integer getFlightId() {
        return flightId;
    }

    public void setFlightId(Integer flightId) {
        this.flightId = flightId;
    }

    public String getFlightNumber() {
        return flightNumber;
    }

    public void setFlightNumber(String flightNumber) {
        this.flightNumber = flightNumber;
    }

    public Integer getStatus() {
        return status;
    }

    public void setStatus(Integer status) {
        this.status = status;
    }
}